task #21 SizeChangeCommand refactoring: methods updateValue() and getCurrentValue() pushed down to subclasses and type speculation removed
